Allow compilation as CMAKE subdirectory
This pull request allows the complete library repo to be included as subdirectory of an existing project.

When I included the repo as submodule, I got a "file not found error" of the muParser.h header file. I traced it back to the use of a specific CMAKE variable when adding the include path. According to https://stackoverflow.com/questions/53195186/difference-between-cmake-current-source-dir-cmake-source-dir-variables, CMAKE_CURRENT_SOURCE_DIR should be used instead of CMAKE_SOURCE_DIR.