add mu::generic_callable_type: centralize the calls to fun_types, bulkfun_types, strfun_types, and multfun_type
muparser used to recast to the concrete function pointer type and
then do the call in open-coded ways in multiple different places.
Add mu::generic_callable_type to do just that in a centralized place,
with a few templates.
The goal is to be able, in future patches, to easily modify
generic_callable_type so that it e.g. optionally includes user data.

Some details about the changes:

* generic_callable_type is created; that's a pod struct with a few
methods, most templated, to cast to the good type and forward the
call.  The template param NbParams is used as a direct transcription
of N in e.g. fun_typeN even if it is redundant with the number of
parameters used, as to not increase the risk of mistake between the
wanted function type and the one used at call sites.

* generic_fun_type tok.Fun.ptr is replaced by generic_callable_type
tok.Fun.cb

* other generic_fun_type (to pass the pointer around) are replaced
with generic_callable_type

* call sites are adapted to use generic_callable_type::call_* methods.

* eventually generic_fun_type is removed and replaced (where not by
generic_callable_type) by erased_fun_type for multiple reasons:
  - generic_callable_type needs to store the function pointer with
   a "fake" type and I want it both to be a pod for the purpose of
   layout and not convert implicitly from or to other function pointer
   types of muParserDeh.h, in particular generic_fun_type was a synonym
   of fun_type0
  - other parties could have their own patches so it is better
   to at least rename what yields a risk of semantic incompatibility

* generic_callable_type also includes a few utility methods to
cover existing parts of the code that compare it (previously
generic_fun_type) with nullptr, do if(ptr), or compare with
&MathImpl<value_type>::UnaryPlus.

I tested the result by launching 'example1' under valgrind, with ASAN,
or with UBSAN, and saw no issue.

The generated Release binary code of the most obviously impacted
functions has been checked and is unchanged from before this patch
(g++ 6.3, Debian 9, amd64)

I expect the library built on common platforms to stay backward
binary compatible for applications using muparser in a "normal" way,
including through the main classes of the C++ API (mu::Parser,
mu::ParserBase), and it should also be possible to extend
generic_callable_type while keeping this compatibility.

Signed-off-by: Guillaume Knispel <guillaume.knispel@hologic.com>
Reviewed-by: Juan Gomez <juan.gomez@hologic.com>